# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Binogi Code Challenge API",
        "description": "Binogi Code Challenge API - OpenAPI Description",
        "contact": {
            "email": "ops@binogi.com"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost:7777"
        }
    ],
    "paths": {
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show user by ID",
                "description": "Show user by ID",
                "operationId": "d3a14968f8b7a072620c75c7c78ed6f8",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserMapper"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user by ID",
                "description": "Update user by ID",
                "operationId": "e29fc231c7937ab09483d35e6e3d1214",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User after the update",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserMapper"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Failed validation of given params"
                    }
                }
            }
        },
        "/api/users/": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show users",
                "description": "Show user",
                "operationId": "index",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Users Details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserMapper"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/store": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "description": "Create a new user",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserMapper"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User cannot be created"
                    },
                    "422": {
                        "description": "Failed validation of given params"
                    }
                }
            }
        },
        "/api/users/nickname/{nickname}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Find user by nickname",
                "description": "Find user by nickname",
                "operationId": "showByNickname",
                "parameters": [
                    {
                        "name": "nickname",
                        "in": "path",
                        "description": "nickname of user that needs to be fetched",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserMapper"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserMapper": {
                "title": "UserMapper",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "description": "email",
                        "type": "string",
                        "example": "JaneDoe@email.com"
                    },
                    "name": {
                        "description": "name",
                        "type": "string",
                        "example": "Jane Doe"
                    }
                },
                "type": "object"
            },
            "UserStoreRequest": {
                "description": "Create new user",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "maxLength": 191,
                        "minLength": 1,
                        "example": "Jane Doe"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "maxLength": 191,
                        "minLength": 1,
                        "example": "JaneDoe@email.com"
                    },
                    "password": {
                        "description": "User Password",
                        "type": "string",
                        "maxLength": 191,
                        "minLength": 1,
                        "example": "correct horse battery staple"
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "description": "User Update Request Body",
                "properties": {
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "maxLength": 191,
                        "minLength": 1,
                        "example": "Jane Doe"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "maxLength": 191,
                        "minLength": 1,
                        "example": "JaneDoe@email.com"
                    },
                    "password": {
                        "description": "User Password",
                        "type": "string",
                        "maxLength": 191,
                        "minLength": 1,
                        "example": "correct horse battery staple"
                    }
                },
                "type": "object"
            }
        }
    }
}